chore(interface-types) Move the `instruction.rs` module in `instructions/mod.rs`.
